# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trends Laravel API",
        "description": "企業の技術コミュニティでの影響力を分析・追跡するAPI",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/rankings/{period}": {
            "get": {
                "tags": [
                    "企業ランキング"
                ],
                "summary": "期間別ランキング取得",
                "description": "指定した期間の企業ランキングを取得します。",
                "operationId": "2a59ef3da07e77296fd7d929cd34076e",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "description": "期間タイプ (1w, 1m, 3m, 6m, 1y, 3y, all)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "1w",
                                "1m",
                                "3m",
                                "6m",
                                "1y",
                                "3y",
                                "all"
                            ],
                            "example": "1m"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "ページ番号",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "1ページあたりの件数（最大100）",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "example": 20
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "ソート項目",
                        "schema": {
                            "type": "string",
                            "default": "rank_position",
                            "example": "rank_position"
                        }
                    },
                    {
                        "name": "sort_order",
                        "in": "query",
                        "description": "ソート順",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "asc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ランキングリスト",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正な期間タイプ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rankings/{period}/top/{limit}": {
            "get": {
                "tags": [
                    "企業ランキング"
                ],
                "summary": "上位N件のランキング取得",
                "description": "指定した期間の上位N件のランキングを取得します。",
                "operationId": "78594b53d7fae9a30415e698a61b925f",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "description": "期間タイプ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "1w",
                                "1m",
                                "3m",
                                "6m",
                                "1y",
                                "3y",
                                "all"
                            ],
                            "example": "1m"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "取得件数 (1-100)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上位ランキング",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "period": {
                                                    "type": "string"
                                                },
                                                "limit": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "不正なパラメータ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}